refactor(centaur-perms): rename from centaur-grants, resource-first CLI
Rename the crate/binary to centaur-perms and restructure the CLI as
<noun> <verb>:

  principals list | show <p> | grant <p> | revoke <p>
  roles      list | show <r> | grant <r> | revoke <r>

principals grant/revoke take any mix of --tool, --role, and --secret
targets; roles grant/revoke assign/remove secrets on a role (the
previously missing role<-secret path). Adds get_role and list_roles
client methods.
