+++ services/api-rs/Cargo.toml
@@ -1,9 +1,9 @@
 [workspace]
 members = [
     "crates/centaur-api-server",
-    "crates/centaur-grants",
     "crates/centaur-iron-control",
     "crates/centaur-iron-proxy",
+    "crates/centaur-perms",
     "crates/centaur-sandbox-core",
     "crates/centaur-sandbox-e2e",
     "crates/centaur-sandbox-agent-k8s",
